Move `ImageCommentsViewController` to production
